Added a parameter to pass in highlight.js in from the renderConfig
Previously, if you wanted to use highlight.js to render the preview, you had to have hljs on the window object. In practice, this is poor and negates many of the
principles behind using Browserify. By injecting it in as a parameter, this means the client can avoid adding things to their global scope if they don't/can't
avoid that.

This is a non-breaking change, only adding a new option to an existing API. The default behaviour is unchanged