Added a "disabled" configuration option along with a complementary setter method
SimpleMDE can now be disabled through initial configuration, or
disabled / re-enabled using a method dynamically. README documentation
has been updated.

In addition to this, there was a typo in the SimpleMDE JavaScript file,
where "Tootlip" was used (consistently). This was fixed as well.

A demo of the functionality can be found at
http://sweltering-verse.surge.sh